requirements.txt
streamlit
numpy
Pillow
sentence-transformers
faiss-cpu
requests
openai
